refactor: 검증 로직 변경 및 테스틐 코드 수정
변경 내용 :
- LineControllerTest 코드 통합테스트로 위치 이동
- IllegalStateException 예외메시지 추가
- Common으로 패키지 구조 변경
- 기존 테스트코드 리팩터링
