+++ src/main/java/subway/step/StationController.java
@@ -1,13 +1,15 @@
-package subway.ui;
+package subway.step;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import subway.application.StationService;
 import subway.dto.StationRequest;
 import subway.dto.StationResponse;
 
 import java.net.URI;
-import java.sql.SQLException;
 
 @RestController
 @RequestMapping("/stations")
@@ -22,10 +24,5 @@ public class StationController {
     public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
         StationResponse station = stationService.saveStation(stationRequest);
         return ResponseEntity.created(URI.create("/stations/" + station.getId())).body(station);
-    }
-
-    @ExceptionHandler(SQLException.class)
-    public ResponseEntity<Void> handleSQLException() {
-        return ResponseEntity.badRequest().build();
     }
 }
