+++ src/test/java/subway/common/step/StationStep.java
@@ -1,4 +1,4 @@
-package subway.ui;
+package subway.common.step;
 
 import io.restassured.path.json.JsonPath;
 import io.restassured.response.ExtractableResponse;
@@ -10,15 +10,18 @@ import static io.restassured.RestAssured.given;
 
 public class StationStep {
 
-    public static Long createStation(final StationRequest stationRequest) {
-        final ExtractableResponse<Response> response = given().log().all()
+    public static ExtractableResponse<Response> createStation(final StationRequest stationRequest) {
+        return given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(stationRequest)
                 .when()
                 .post("/stations")
                 .then().log().all()
                 .extract();
+    }
 
+    public static Long createStationAndGetId(final StationRequest stationRequest) {
+        final ExtractableResponse<Response> response = createStation(stationRequest);
         final JsonPath jsonPath = response.jsonPath();
         return jsonPath.getLong("id");
     }
