+++ src/test/java/subway/dao/LineDaoTest.java
@@ -15,8 +15,8 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
-import static subway.fixture.EntityFixture.이호선_초록색_Entity;
-import static subway.fixture.EntityFixture.일호선_남색_Entity;
+import static subway.common.fixture.EntityFixture.이호선_초록색_Entity;
+import static subway.common.fixture.EntityFixture.일호선_남색_Entity;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
