+++ src/test/java/subway/dao/SectionDaoTest.java
@@ -14,8 +14,8 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
-import static subway.fixture.EntityFixture.일호선_남색_Entity;
-import static subway.fixture.EntityFixture.후추_디노_Entity;
+import static subway.common.fixture.EntityFixture.일호선_남색_Entity;
+import static subway.common.fixture.EntityFixture.후추_디노_Entity;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
