+++ src/test/java/subway/dao/StationDaoTest.java
@@ -15,8 +15,8 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
-import static subway.fixture.EntityFixture.디노_Entity;
-import static subway.fixture.EntityFixture.후추_Entity;
+import static subway.common.fixture.EntityFixture.디노_Entity;
+import static subway.common.fixture.EntityFixture.후추_Entity;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
