+++ src/test/java/subway/domain/SectionTest.java
@@ -10,7 +10,7 @@ import subway.domain.section.Section;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
-import static subway.fixture.DomainFixture.*;
+import static subway.common.fixture.DomainFixture.*;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
